Move counter and errer to subpackage.
This should clean up the docs.
