+++ basic_test.go
@@ -5,6 +5,8 @@ import (
 	"log"
 	"testing"
 	"time"
+
+	blog "github.com/realzeitmedia/bubbles/log"
 )
 
 func init() {
@@ -200,3 +202,33 @@ func TestShutdownTimeout(t *testing.T) {
 		t.Errorf("Wrong pending object returned")
 	}
 }
+
+type val struct{ C, T int }
+
+type count struct {
+	Sends      int
+	Retries    int
+	Errors     int
+	SendTotals val
+	Troubles   int
+}
+
+func (c *count) Actions(s, r, e int) {
+	c.Sends += s
+	c.Retries += r
+	c.Errors += e
+}
+
+func (c *count) SendTotal(l int) {
+	c.SendTotals.C++
+	c.SendTotals.T += l
+}
+
+func (c *count) Trouble() {
+	c.Troubles++
+}
+
+func (c *count) BatchTime(time.Duration) {
+}
+
+var _ blog.Counter = &count{}
